removed legacy code dependency
